# gateway: optimize http route generation for gateways