--- a/pilot/pkg/xds/testdata/benchmarks/gateways.yaml
+++ b/pilot/pkg/xds/testdata/benchmarks/gateways.yaml
@@ -67,6 +67,18 @@ spec:
     - random-c-{{$i}}.host.example
     tls:
       mode: ISTIO_MUTUAL
+  - port:
+      number: 443
+      name: https-simple
+      protocol: HTTPS
+    hosts:
+    - random-a-{{$i}}.host.example
+    - random-b-{{$i}}.host.example
+    - random-c-{{$i}}.host.example
+    tls:
+      mode: SIMPLE
+      serverCertificate: /etc/certs/servercert.pem
+      privateKey: /etc/certs/privatekey.pem
 ---
 apiVersion: networking.istio.io/v1alpha3
 kind: ServiceEntry