--- a/pilot/pkg/xds/testdata/benchmarks/gateways.yaml
+++ b/pilot/pkg/xds/testdata/benchmarks/gateways.yaml
@@ -69,7 +69,67 @@ spec:
       mode: ISTIO_MUTUAL
   - port:
       number: 443
-      name: https-simple
+      name: https-simple-1
+      protocol: HTTPS
+    hosts:
+    - random-a-{{$i}}.host.example
+    - random-b-{{$i}}.host.example
+    - random-c-{{$i}}.host.example
+    tls:
+      mode: SIMPLE
+      serverCertificate: /etc/certs/servercert.pem
+      privateKey: /etc/certs/privatekey.pem
+  - port:
+      number: 443
+      name: https-simple-2
+      protocol: HTTPS
+    hosts:
+    - random-a-{{$i}}.host.example
+    - random-b-{{$i}}.host.example
+    - random-c-{{$i}}.host.example
+    tls:
+      mode: SIMPLE
+      serverCertificate: /etc/certs/servercert.pem
+      privateKey: /etc/certs/privatekey.pem
+  - port:
+      number: 443
+      name: https-simple-3
+      protocol: HTTPS
+    hosts:
+    - random-a-{{$i}}.host.example
+    - random-b-{{$i}}.host.example
+    - random-c-{{$i}}.host.example
+    tls:
+      mode: SIMPLE
+      serverCertificate: /etc/certs/servercert.pem
+      privateKey: /etc/certs/privatekey.pem
+  - port:
+      number: 443
+      name: https-simple-4
+      protocol: HTTPS
+    hosts:
+    - random-a-{{$i}}.host.example
+    - random-b-{{$i}}.host.example
+    - random-c-{{$i}}.host.example
+    tls:
+      mode: SIMPLE
+      serverCertificate: /etc/certs/servercert.pem
+      privateKey: /etc/certs/privatekey.pem
+  - port:
+      number: 443
+      name: https-simple-5
+      protocol: HTTPS
+    hosts:
+    - random-a-{{$i}}.host.example
+    - random-b-{{$i}}.host.example
+    - random-c-{{$i}}.host.example
+    tls:
+      mode: SIMPLE
+      serverCertificate: /etc/certs/servercert.pem
+      privateKey: /etc/certs/privatekey.pem
+  - port:
+      number: 443
+      name: https-simple-6
       protocol: HTTPS
     hosts:
     - random-a-{{$i}}.host.example