@ramaraochavali 2020-08-03
Benchmark results before and after

```
name                                old time/op        new time/op        delta
RouteGeneration/gateways-12                479ms ±10%         490ms ±10%     ~     (p=1.000 n=5+5)
RouteGeneration/gateways-shared-12         2.34s ± 7%         1.69s ± 6%  -27.62%  (p=0.008 n=5+5)
RouteGeneration/empty-12                  5.91ms ±10%        5.93ms ±14%     ~     (p=1.000 n=5+5)
RouteGeneration/tls-12                    5.55ms ± 2%        5.87ms ± 9%     ~     (p=0.190 n=4+5)
RouteGeneration/telemetry-12              6.27ms ±18%        5.96ms ±11%     ~     (p=1.000 n=5+5)
RouteGeneration/virtualservice-12         17.5ms ±10%        15.3ms ±12%  -12.15%  (p=0.032 n=5+5)

name                                old kb/msg         new kb/msg         delta
RouteGeneration/gateways-12                2.73k ± 0%         2.73k ± 0%     ~     (all equal)
RouteGeneration/gateways-shared-12         12.6k ± 0%         12.3k ± 0%   -2.69%  (p=0.008 n=5+5)
RouteGeneration/empty-12                    32.0 ± 0%          32.0 ± 0%     ~     (all equal)
RouteGeneration/tls-12                      32.0 ± 0%          32.0 ± 0%     ~     (all equal)
RouteGeneration/telemetry-12                32.0 ± 0%          32.0 ± 0%     ~     (all equal)
RouteGeneration/virtualservice-12           80.9 ± 0%          80.9 ± 0%     ~     (all equal)

name                                old resources/msg  new resources/msg  delta
RouteGeneration/gateways-12                1.00k ± 0%         1.00k ± 0%     ~     (all equal)
RouteGeneration/gateways-shared-12          2.00 ± 0%          2.00 ± 0%     ~     (all equal)
RouteGeneration/empty-12                    2.00 ± 0%          2.00 ± 0%     ~     (all equal)
RouteGeneration/tls-12                      2.00 ± 0%          2.00 ± 0%     ~     (all equal)
RouteGeneration/telemetry-12                2.00 ± 0%          2.00 ± 0%     ~     (all equal)
RouteGeneration/virtualservice-12           1.00 ± 0%          1.00 ± 0%     ~     (all equal)

name                                old alloc/op       new alloc/op       delta
RouteGeneration/gateways-12               15.7MB ± 0%        15.9MB ± 0%   +1.28%  (p=0.008 n=5+5)
RouteGeneration/gateways-shared-12        67.8MB ± 0%        37.4MB ± 0%  -44.91%  (p=0.000 n=4+5)
RouteGeneration/empty-12                   568kB ± 0%         568kB ± 0%     ~     (p=0.516 n=5+5)
RouteGeneration/tls-12                     567kB ± 0%         567kB ± 0%     ~     (p=0.190 n=5+4)
RouteGeneration/telemetry-12               568kB ± 0%         568kB ± 0%     ~     (p=0.548 n=5+5)
RouteGeneration/virtualservice-12          866kB ± 0%         866kB ± 0%     ~     (p=0.937 n=5+5)

name                                old allocs/op      new allocs/op      delta
RouteGeneration/gateways-12                 205k ± 0%          205k ± 0%   +0.01%  (p=0.008 n=5+5)
RouteGeneration/gateways-shared-12          902k ± 0%          511k ± 0%  -43.41%  (p=0.000 n=4+5)
RouteGeneration/empty-12                   5.82k ± 1%         5.81k ± 1%     ~     (p=0.762 n=5+5)
RouteGeneration/tls-12                     5.79k ± 0%         5.79k ± 0%     ~     (p=0.810 n=5+4)
RouteGeneration/telemetry-12               5.81k ± 0%         5.81k ± 0%     ~     (p=1.000 n=5+5)
RouteGeneration/virtualservice-12          11.0k ± 0%         11.0k ± 0%     ~     (p=0.429 n=4+4)
```