@ramaraochavali 2020-08-03
```
name                                old time/op        new time/op        delta
RouteGeneration/gateways-12                568ms ±32%         466ms ± 7%     ~     (p=0.222 n=5+5)
RouteGeneration/gateways-shared-12         2.66s ±33%         2.20s ±10%     ~     (p=0.421 n=5+5)
RouteGeneration/empty-12                  6.79ms ±32%        5.75ms ±10%     ~     (p=0.421 n=5+5)
RouteGeneration/tls-12                    6.92ms ±33%        5.43ms ± 3%  -21.51%  (p=0.016 n=5+5)
RouteGeneration/telemetry-12              6.98ms ±28%        5.79ms ±14%     ~     (p=0.151 n=5+5)
RouteGeneration/virtualservice-12         19.0ms ±28%        16.1ms ±22%     ~     (p=0.151 n=5+5)
```